I-JEPAのインターフェース整備
