+++ ami/models/i_jepa.py
@@ -1,689 +1,65 @@
 # Ref: https://github.com/facebookresearch/ijepa
 
 import math
-from functools import partial
 import numpy as np
 
 import torch
 import torch.nn as nn
 
+import dataclasses
 
-def _no_grad_trunc_normal_(
-    tensor: torch.Tensor, mean: float, std: float, a: float, b: float
-) -> torch.Tensor:
-    # Cut & paste from PyTorch official master until it's in a few official releases - RW
-    # Method based on https://people.sc.fsu.edu/~jburkardt/presentations/truncated_normal.pdf
-    def norm_cdf(x: float) -> float:
-        # Computes standard normal cumulative distribution function
-        return (1.0 + math.erf(x / math.sqrt(2.0))) / 2.0
-
-    with torch.no_grad():
-        # Values are generated by using a truncated uniform distribution and
-        # then using the inverse CDF for the normal distribution.
-        # Get upper and lower cdf values
-        l = norm_cdf((a - mean) / std)
-        u = norm_cdf((b - mean) / std)
-
-        # Uniformly fill tensor with values from [l, u], then translate to
-        # [2l-1, 2u-1].
-        tensor.uniform_(2 * l - 1, 2 * u - 1)
-
-        # Use inverse cdf transform for normal distribution to get truncated
-        # standard normal
-        tensor.erfinv_()
-
-        # Transform to proper mean, std
-        tensor.mul_(std * math.sqrt(2.0))
-        tensor.add_(mean)
-
-        # Clamp to ensure it's in the proper range
-        tensor.clamp_(min=a, max=b)
-        return tensor
+IJEPAEncoderSize = TypeVar(
+    "IJEPAEncoderSize", Literal["tiny", "small", "base", "large", "huge", "giant"]
+)
 
 
-def trunc_normal_(
-    tensor: torch.Tensor,
-    mean: float = 0.0,
-    std: float = 1.0,
-    a: float = -2.0,
-    b: float = 2.0,
-) -> torch.Tensor:
-    return _no_grad_trunc_normal_(tensor, mean, std, a, b)
+@dataclass
+class IJEPAEncoderSizeParams:
+    embed_dim: int
+    depth: int
+    num_heads: int
+    mlp_ratio: int
 
 
-def repeat_interleave_batch(x: torch.Tensor, B: int, repeat: int) -> torch.Tensor:
-    N = len(x) // B
-    x = torch.cat(
-        [
-            torch.cat([x[i * B : (i + 1) * B] for _ in range(repeat)], dim=0)
-            for i in range(N)
-        ],
-        dim=0,
-    )
-    return x
+encoder_sizes: Dict[IJEPAEncoderSize, IJEPAEncoderSizeParams] = {
+    "tiny": IJEPAEncoderSizeParams(embed_dim=192, depth=12, num_heads=3, mlp_ratio=4),
+    "small": IJEPAEncoderSizeParams(embed_dim=384, depth=12, num_heads=6, mlp_ratio=4),
+    "base": IJEPAEncoderSizeParams(embed_dim=768, depth=12, num_heads=12, mlp_ratio=4),
+    "large": IJEPAEncoderSizeParams(
+        embed_dim=1024, depth=24, num_heads=16, mlp_ratio=4
+    ),
+    "huge": IJEPAEncoderSizeParams(embed_dim=1280, depth=32, num_heads=16, mlp_ratio=4),
+    "giant": IJEPAEncoderSizeParams(
+        embed_dim=1408, depth=40, num_heads=16, mlp_ratio=48 / 11
+    ),
+}
 
 
-def apply_masks(x: torch.Tensor, masks: List[torch.Tensor]) -> torch.Tensor:
-    """
-    :param x: tensor of shape [B (batch-size), N (num-patches), D (feature-dim)]
-    :param masks: list of tensors containing indices of patches in [N] to keep
-    """
-    all_x = []
-    for m in masks:
-        mask_keep = m.unsqueeze(-1).repeat(1, 1, x.size(-1))
-        all_x += [torch.gather(x, dim=1, index=mask_keep)]
-    return torch.cat(all_x, dim=0)
-
-
-def get_2d_sincos_pos_embed(
-    embed_dim: int, grid_size: int, cls_token: bool = False
-) -> np.ndarray:
-    """
-    grid_size: int of the grid height and width
-    return:
-    pos_embed: [grid_size*grid_size, embed_dim] or [1+grid_size*grid_size, embed_dim] (w/ or w/o cls_token)
-    """
-    grid_h = np.arange(grid_size, dtype=float)
-    grid_w = np.arange(grid_size, dtype=float)
-    grid = np.meshgrid(grid_w, grid_h)  # here w goes first
-    grid = np.stack(grid, axis=0)
-
-    grid = grid.reshape([2, 1, grid_size, grid_size])
-    pos_embed = get_2d_sincos_pos_embed_from_grid(embed_dim, grid)
-    if cls_token:
-        pos_embed = np.concatenate([np.zeros([1, embed_dim]), pos_embed], axis=0)
-    return pos_embed
-
-
-def get_2d_sincos_pos_embed_from_grid(
-    embed_dim: int, grid: Tuple[int, int]
-) -> np.ndarray:
-    assert embed_dim % 2 == 0
-
-    # use half of dimensions to encode grid_h
-    emb_h = get_1d_sincos_pos_embed_from_grid(embed_dim // 2, grid[0])  # (H*W, D/2)
-    emb_w = get_1d_sincos_pos_embed_from_grid(embed_dim // 2, grid[1])  # (H*W, D/2)
-
-    emb = np.concatenate([emb_h, emb_w], axis=1)  # (H*W, D)
-    return emb
-
-
-def get_1d_sincos_pos_embed(
-    embed_dim: int, grid_size: int, cls_token=False
-) -> np.ndarray:
-    """
-    grid_size: int of the grid length
-    return:
-    pos_embed: [grid_size, embed_dim] or [1+grid_size, embed_dim] (w/ or w/o cls_token)
-    """
-    grid = np.arange(grid_size, dtype=float)
-    pos_embed = get_1d_sincos_pos_embed_from_grid(embed_dim, grid)
-    if cls_token:
-        pos_embed = np.concatenate([np.zeros([1, embed_dim]), pos_embed], axis=0)
-    return pos_embed
-
-
-def get_1d_sincos_pos_embed_from_grid(embed_dim: int, pos: np.ndarray) -> np.ndarray:
-    """
-    embed_dim: output dimension for each position
-    pos: a list of positions to be encoded: size (M,)
-    out: (M, D)
-    """
-    assert embed_dim % 2 == 0
-    omega = np.arange(embed_dim // 2, dtype=float)
-    omega /= embed_dim / 2.0
-    omega = 1.0 / 10000**omega  # (D/2,)
-
-    pos = pos.reshape(-1)  # (M,)
-    out = np.einsum("m,d->md", pos, omega)  # (M, D/2), outer product
-
-    emb_sin = np.sin(out)  # (M, D/2)
-    emb_cos = np.cos(out)  # (M, D/2)
-
-    emb = np.concatenate([emb_sin, emb_cos], axis=1)  # (M, D)
-    return emb
-
-
-def drop_path(
-    x: torch.Tensor, drop_prob: float = 0.0, training: bool = False
-) -> torch.Tensor:
-    if drop_prob == 0.0 or not training:
-        return x
-    keep_prob = 1 - drop_prob
-    shape = (x.shape[0],) + (1,) * (
-        x.ndim - 1
-    )  # work with diff dim tensors, not just 2D ConvNets
-    random_tensor = keep_prob + torch.rand(shape, dtype=x.dtype, device=x.device)
-    random_tensor.floor_()  # binarize
-    output = x.div(keep_prob) * random_tensor
-    return output
-
-
-class DropPath(nn.Module):
-    """Drop paths (Stochastic Depth) per sample  (when applied in main path of residual blocks)."""
-
-    def __init__(self, drop_prob: float):
-        super(DropPath, self).__init__()
-        assert drop_path > 0.0
-        self.drop_prob = drop_prob
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return drop_path(x, self.drop_prob, self.training)
-
-
-class MLP(nn.Module):
+class IJEPAEncoder(VisionTransformer):
     def __init__(
-        self,
-        in_features: int,
-        hidden_features: int,
-        out_features: int,
-        drop: float = 0.0,
+        self, img_size: int, patch_size: int, model_size: IJEPAEncoderSize
     ) -> None:
-        super().__init__()
-        self.fc1 = nn.Linear(in_features, hidden_features)
-        self.act = nn.GELU()
-        self.fc2 = nn.Linear(hidden_features, out_features)
-        self.drop = nn.Dropout(drop)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = self.fc1(x)
-        x = self.act(x)
-        x = self.drop(x)
-        x = self.fc2(x)
-        x = self.drop(x)
-        return x
-
-
-class Attention(nn.Module):
-    def __init__(
-        self,
-        dim: int,
-        num_heads: int = 8,
-        qkv_bias: bool = False,
-        qk_scale: Optional[float] = None,
-        attn_drop: float = 0.0,
-        proj_drop: float = 0.0,
-    ) -> None:
-        super().__init__()
-        self.num_heads = num_heads
-        head_dim = dim // num_heads
-        self.scale = qk_scale or head_dim**-0.5
-
-        self.qkv = nn.Linear(dim, dim * 3, bias=qkv_bias)
-        self.attn_drop = nn.Dropout(attn_drop)
-        self.proj = nn.Linear(dim, dim)
-        self.proj_drop = nn.Dropout(proj_drop)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        B, N, C = x.shape
-        qkv = (
-            self.qkv(x)
-            .reshape(B, N, 3, self.num_heads, C // self.num_heads)
-            .permute(2, 0, 3, 1, 4)
-        )
-        q, k, v = qkv[0], qkv[1], qkv[2]
-
-        attn = (q @ k.transpose(-2, -1)) * self.scale
-        attn = attn.softmax(dim=-1)
-        attn = self.attn_drop(attn)
-
-        x = (attn @ v).transpose(1, 2).reshape(B, N, C)
-        x = self.proj(x)
-        x = self.proj_drop(x)
-        return x, attn
-
-
-class Block(nn.Module):
-    def __init__(
-        self,
-        dim: int,
-        num_heads: int,
-        mlp_ratio: float = 4.0,
-        qkv_bias: bool = False,
-        qk_scale: Optional[float] = None,
-        drop: float = 0.0,
-        attn_drop: float = 0.0,
-        drop_path: float = 0.0,
-    ) -> None:
-        super().__init__()
-        self.norm1 = nn.LayerNorm(dim, eps=1e-6)
-        self.attn = Attention(
-            dim,
-            num_heads=num_heads,
-            qkv_bias=qkv_bias,
-            qk_scale=qk_scale,
-            attn_drop=attn_drop,
-            proj_drop=drop,
-        )
-        self.drop_path = DropPath(drop_path) if drop_path > 0.0 else nn.Identity()
-        self.norm2 = nn.LayerNorm(dim, eps=1e-6)
-        mlp_hidden_dim = int(dim * mlp_ratio)
-        self.mlp = MLP(
-            in_features=dim,
-            hidden_features=mlp_hidden_dim,
-            out_features=dim,
-            drop=drop,
-        )
-
-    def forward(self, x: torch.Tensor, return_attention: bool = False) -> torch.Tensor:
-        y, attn = self.attn(self.norm1(x))
-        if return_attention:
-            return attn
-        x = x + self.drop_path(y)
-        x = x + self.drop_path(self.mlp(self.norm2(x)))
-        return x
-
-
-class PatchEmbed(nn.Module):
-    """Image to Patch Embedding"""
-
-    def __init__(
-        self,
-        img_size: int = 224,
-        patch_size: int = 16,
-        in_chans: int = 3,
-        embed_dim: int = 768,
-    ) -> None:
-        super().__init__()
-        num_patches = (img_size // patch_size) * (img_size // patch_size)
-        self.img_size = img_size
-        self.patch_size = patch_size
-        self.num_patches = num_patches
-
-        self.proj = nn.Conv2d(
-            in_chans, embed_dim, kernel_size=patch_size, stride=patch_size
-        )
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        B, C, H, W = x.shape
-        x = self.proj(x).flatten(2).transpose(1, 2)
-        return x
-
-
-class ConvEmbed(nn.Module):
-    """
-    3x3 Convolution stems for ViT following ViTC models
-    """
-
-    def __init__(
-        self,
-        channels: int,
-        strides: int,
-        img_size: int = 224,
-        in_chans: int = 3,
-        batch_norm: bool = True,
-    ) -> None:
-        super().__init__()
-        # Build the stems
-        stem = []
-        channels = [in_chans] + channels
-        for i in range(len(channels) - 2):
-            stem += [
-                nn.Conv2d(
-                    channels[i],
-                    channels[i + 1],
-                    kernel_size=3,
-                    stride=strides[i],
-                    padding=1,
-                    bias=(not batch_norm),
-                )
-            ]
-            if batch_norm:
-                stem += [nn.BatchNorm2d(channels[i + 1])]
-            stem += [nn.ReLU(inplace=True)]
-        stem += [
-            nn.Conv2d(channels[-2], channels[-1], kernel_size=1, stride=strides[-1])
-        ]
-        self.stem = nn.Sequential(*stem)
-
-        # Comptute the number of patches
-        stride_prod = int(np.prod(strides))
-        self.num_patches = (img_size[0] // stride_prod) ** 2
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        p = self.stem(x)
-        return p.flatten(2).transpose(1, 2)
-
-
-class VisionTransformerPredictor(nn.Module):
-    """Vision Transformer"""
-
-    def __init__(
-        self,
-        num_patches: int,
-        embed_dim: int = 768,
-        predictor_embed_dim: int = 384,
-        depth: int = 6,
-        num_heads: int = 12,
-        mlp_ratio: float = 4.0,
-        qkv_bias: bool = True,
-        qk_scale: Optional[float] = None,
-        drop_rate: float = 0.0,
-        attn_drop_rate: float = 0.0,
-        drop_path_rate: float = 0.0,
-        init_std: float = 0.02,
-        **kwargs
-    ) -> None:
-        super().__init__()
-        self.predictor_embed = nn.Linear(embed_dim, predictor_embed_dim, bias=True)
-        self.mask_token = nn.Parameter(torch.zeros(1, 1, predictor_embed_dim))
-        dpr = [
-            x.item() for x in torch.linspace(0, drop_path_rate, depth)
-        ]  # stochastic depth decay rule
-        # --
-        self.predictor_pos_embed = nn.Parameter(
-            torch.zeros(1, num_patches, predictor_embed_dim), requires_grad=False
-        )
-        predictor_pos_embed = get_2d_sincos_pos_embed(
-            self.predictor_pos_embed.shape[-1], int(num_patches**0.5), cls_token=False
-        )
-        self.predictor_pos_embed.data.copy_(
-            torch.from_numpy(predictor_pos_embed).float().unsqueeze(0)
-        )
-        # --
-        self.predictor_blocks = nn.ModuleList(
-            [
-                Block(
-                    dim=predictor_embed_dim,
-                    num_heads=num_heads,
-                    mlp_ratio=mlp_ratio,
-                    qkv_bias=qkv_bias,
-                    qk_scale=qk_scale,
-                    drop=drop_rate,
-                    attn_drop=attn_drop_rate,
-                    drop_path=dpr[i],
-                    norm_layer=norm_layer,
-                )
-                for i in range(depth)
-            ]
-        )
-        self.predictor_norm = nn.LayerNorm(predictor_embed_dim, eps=1e-6)
-        self.predictor_proj = nn.Linear(predictor_embed_dim, embed_dim, bias=True)
-        # ------
-        self.init_std = init_std
-        trunc_normal_(self.mask_token, std=self.init_std)
-        self.apply(self._init_weights)
-        self.fix_init_weight()
-
-    def fix_init_weight(self) -> None:
-        def rescale(param, layer_id):
-            param.div_(math.sqrt(2.0 * layer_id))
-
-        for layer_id, layer in enumerate(self.predictor_blocks):
-            rescale(layer.attn.proj.weight.data, layer_id + 1)
-            rescale(layer.mlp.fc2.weight.data, layer_id + 1)
-
-    def _init_weights(self, m: nn.Module) -> None:
-        if isinstance(m, nn.Linear):
-            trunc_normal_(m.weight, std=self.init_std)
-            if isinstance(m, nn.Linear) and m.bias is not None:
-                nn.init.constant_(m.bias, 0)
-        elif isinstance(m, nn.LayerNorm):
-            nn.init.constant_(m.bias, 0)
-            nn.init.constant_(m.weight, 1.0)
-        elif isinstance(m, nn.Conv2d):
-            trunc_normal_(m.weight, std=self.init_std)
-            if m.bias is not None:
-                nn.init.constant_(m.bias, 0)
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        masks_x: List[torch.Tensor],
-        masks: List[torch.Tensor],
-    ) -> torch.Tensor:
-        assert (masks is not None) and (
-            masks_x is not None
-        ), "Cannot run predictor without mask indices"
-
-        # -- Batch Size
-        B = len(x) // len(masks_x)
-
-        # -- map from encoder-dim to pedictor-dim
-        x = self.predictor_embed(x)
-
-        # -- add positional embedding to x tokens
-        x_pos_embed = self.predictor_pos_embed.repeat(B, 1, 1)
-        x += apply_masks(x_pos_embed, masks_x)
-
-        _, N_ctxt, D = x.shape
-
-        # -- concat mask tokens to x
-        pos_embs = self.predictor_pos_embed.repeat(B, 1, 1)
-        pos_embs = apply_masks(pos_embs, masks)
-        pos_embs = repeat_interleave_batch(pos_embs, B, repeat=len(masks_x))
-        # --
-        pred_tokens = self.mask_token.repeat(pos_embs.size(0), pos_embs.size(1), 1)
-        # --
-        pred_tokens += pos_embs
-        x = x.repeat(len(masks), 1, 1)
-        x = torch.cat([x, pred_tokens], dim=1)
-
-        # -- fwd prop
-        for blk in self.predictor_blocks:
-            x = blk(x)
-        x = self.predictor_norm(x)
-
-        # -- return preds for mask tokens
-        x = x[:, N_ctxt:]
-        x = self.predictor_proj(x)
-
-        return x
-
-
-class VisionTransformer(nn.Module):
-    """Vision Transformer"""
-
-    def __init__(
-        self,
-        img_size: int = 224,
-        patch_size: int = 16,
-        in_chans: int = 3,
-        embed_dim: int = 768,
-        predictor_embed_dim: int = 384,
-        depth: int = 12,
-        predictor_depth: int = 12,
-        num_heads: int = 12,
-        mlp_ratio: float = 4.0,
-        qkv_bias: bool = True,
-        qk_scale: Optional[float] = None,
-        drop_rate: float = 0.0,
-        attn_drop_rate: float = 0.0,
-        drop_path_rate: float = 0.0,
-        init_std: float = 0.02,
-        **kwargs
-    ) -> None:
-        super().__init__()
-        self.num_features = self.embed_dim = embed_dim
-        self.num_heads = num_heads
-        # --
-        self.patch_embed = PatchEmbed(
+        super().__init__(
             img_size=img_size,
             patch_size=patch_size,
-            in_chans=in_chans,
+            **encoder_sizes[model_size].asdict(),
+        )
+        assert img_size % patch_size == 0
+
+
+class IJEPAPredictor(VisionTransformerPredictor):
+    def __init__(
+        self,
+        patch_size: int,  # encoder.patch_embed.num_patches,
+        embed_dim: int,  # encoder.embed_dim,
+        num_heads: int,  # encoder.num_heads
+        predictor_embed_dim: int = 384,
+        depth: int = 12,
+    ) -> None:
+        super().__init__(
+            patch_size=patch_size,
             embed_dim=embed_dim,
+            predictor_embed_dim=predictor_embed_dim,
+            depth=pred_depth,
+            num_heads=num_heads,
         )
-        num_patches = self.patch_embed.num_patches
-        # --
-        self.pos_embed = nn.Parameter(
-            torch.zeros(1, num_patches, embed_dim), requires_grad=False
-        )
-        pos_embed = get_2d_sincos_pos_embed(
-            self.pos_embed.shape[-1],
-            int(self.patch_embed.num_patches**0.5),
-            cls_token=False,
-        )
-        self.pos_embed.data.copy_(torch.from_numpy(pos_embed).float().unsqueeze(0))
-        # --
-        dpr = [
-            x.item() for x in torch.linspace(0, drop_path_rate, depth)
-        ]  # stochastic depth decay rule
-        self.blocks = nn.ModuleList(
-            [
-                Block(
-                    dim=embed_dim,
-                    num_heads=num_heads,
-                    mlp_ratio=mlp_ratio,
-                    qkv_bias=qkv_bias,
-                    qk_scale=qk_scale,
-                    drop=drop_rate,
-                    attn_drop=attn_drop_rate,
-                    drop_path=dpr[i],
-                    norm_layer=norm_layer,
-                )
-                for i in range(depth)
-            ]
-        )
-        self.norm = nn.LayerNorm(embed_dim, eps=1e-6)
-        # ------
-        self.init_std = init_std
-        self.apply(self._init_weights)
-        self.fix_init_weight()
-
-    def fix_init_weight(self) -> None:
-        def rescale(param, layer_id):
-            param.div_(math.sqrt(2.0 * layer_id))
-
-        for layer_id, layer in enumerate(self.blocks):
-            rescale(layer.attn.proj.weight.data, layer_id + 1)
-            rescale(layer.mlp.fc2.weight.data, layer_id + 1)
-
-    def _init_weights(self, m: nn.Module) -> None:
-        if isinstance(m, nn.Linear):
-            trunc_normal_(m.weight, std=self.init_std)
-            if isinstance(m, nn.Linear) and m.bias is not None:
-                nn.init.constant_(m.bias, 0)
-        elif isinstance(m, nn.LayerNorm):
-            nn.init.constant_(m.bias, 0)
-            nn.init.constant_(m.weight, 1.0)
-        elif isinstance(m, nn.Conv2d):
-            trunc_normal_(m.weight, std=self.init_std)
-            if m.bias is not None:
-                nn.init.constant_(m.bias, 0)
-
-    def forward(
-        self, x: torch.Tensor, masks: Optional[List[torch.Tensor]] = None
-    ) -> torch.Tensor:
-        # -- patchify x
-        x = self.patch_embed(x)
-        B, N, D = x.shape
-
-        # -- add positional embedding to x
-        pos_embed = self.interpolate_pos_encoding(x, self.pos_embed)
-        x = x + pos_embed
-
-        # -- mask x
-        if masks is not None:
-            x = apply_masks(x, masks)
-
-        # -- fwd prop
-        for i, blk in enumerate(self.blocks):
-            x = blk(x)
-
-        if self.norm is not None:
-            x = self.norm(x)
-
-        return x
-
-    def interpolate_pos_encoding(
-        self, x: torch.Tensor, pos_embed: torch.Tensor
-    ) -> torch.Tensor:
-        npatch = x.shape[1] - 1
-        N = pos_embed.shape[1] - 1
-        if npatch == N:
-            return pos_embed
-        class_emb = pos_embed[:, 0]
-        pos_embed = pos_embed[:, 1:]
-        dim = x.shape[-1]
-        pos_embed = nn.functional.interpolate(
-            pos_embed.reshape(1, int(math.sqrt(N)), int(math.sqrt(N)), dim).permute(
-                0, 3, 1, 2
-            ),
-            scale_factor=math.sqrt(npatch / N),
-            mode="bicubic",
-        )
-        pos_embed = pos_embed.permute(0, 2, 3, 1).view(1, -1, dim)
-        return torch.cat((class_emb.unsqueeze(0), pos_embed), dim=1)
-
-
-def i_jepa_encoder_predictor(**kwargs) -> VisionTransformerPredictor:
-    model = VisionTransformerPredictor(mlp_ratio=4, qkv_bias=True, **kwargs)
-    return model
-
-
-def i_jepa_encoder_tiny(patch_size: int = 16, **kwargs) -> VisionTransformer:
-    model = VisionTransformer(
-        patch_size=patch_size,
-        embed_dim=192,
-        depth=12,
-        num_heads=3,
-        mlp_ratio=4,
-        qkv_bias=True,
-        **kwargs
-    )
-    return model
-
-
-def i_jepa_encoder_small(patch_size: int = 16, **kwargs) -> VisionTransformer:
-    model = VisionTransformer(
-        patch_size=patch_size,
-        embed_dim=384,
-        depth=12,
-        num_heads=6,
-        mlp_ratio=4,
-        qkv_bias=True,
-        **kwargs
-    )
-    return model
-
-
-def i_jepa_encoder_base(patch_size: int = 16, **kwargs) -> VisionTransformer:
-    model = VisionTransformer(
-        patch_size=patch_size,
-        embed_dim=768,
-        depth=12,
-        num_heads=12,
-        mlp_ratio=4,
-        qkv_bias=True,
-        **kwargs
-    )
-    return model
-
-
-def i_jepa_encoder_large(patch_size: int = 16, **kwargs) -> VisionTransformer:
-    model = VisionTransformer(
-        patch_size=patch_size,
-        embed_dim=1024,
-        depth=24,
-        num_heads=16,
-        mlp_ratio=4,
-        qkv_bias=True,
-        **kwargs
-    )
-    return model
-
-
-def i_jepa_encoder_huge(patch_size: int = 16, **kwargs) -> VisionTransformer:
-    model = VisionTransformer(
-        patch_size=patch_size,
-        embed_dim=1280,
-        depth=32,
-        num_heads=16,
-        mlp_ratio=4,
-        qkv_bias=True,
-        **kwargs
-    )
-    return model
-
-
-def i_jepa_encoder_giant(patch_size: int = 16, **kwargs) -> VisionTransformer:
-    model = VisionTransformer(
-        patch_size=patch_size,
-        embed_dim=1408,
-        depth=40,
-        num_heads=16,
-        mlp_ratio=48 / 11,
-        qkv_bias=True,
-        **kwargs
-    )
-    return model
